Reorganize server package structure
