+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/fetcher/resources/ContestListGson.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.resources;
+package com.gustavo.competitiveprogrammingapp.cfApi.fetcher.resources;
 
 data class ContestListGson(
     val id: Int?,
