+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/information/Processor.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.processors
+package com.gustavo.competitiveprogrammingapp.cfApi.information
 
 /* Processors update the server database given API structured data */
 interface Processor {
